Move sync processing to bull queue (#1411)
