+++ creator-node/src/ImageProcessingQueue.js
@@ -31,10 +31,15 @@ class ImageProcessingQueue {
       }
     )
 
+    /**
+     * Queue will process tasks concurrently if provided a concurrency number and a
+     *    path to file containing job processor function
+     * https://github.com/OptimalBits/bull/tree/013c51942e559517c57a117c27a550a0fb583aa8#separate-processes
+     */
     this.queue.process(
-      PROCESS_NAMES.resizeImage,
-      MAX_CONCURRENCY,
-      `${__dirname}/resizeImage.js`
+      PROCESS_NAMES.resizeImage /** job processor name */,
+      MAX_CONCURRENCY /** job processor concurrency */,
+      `${__dirname}/resizeImage.js` /** path to job processor function */
     )
 
     this.logStatus = this.logStatus.bind(this)
