+++ creator-node/src/config.js
@@ -273,12 +273,18 @@ const config = convict({
     default: -1
   },
 
-  // NodeSync filesave concurrency limit
+  // Sync configs
   nodeSyncFileSaveMaxConcurrency: {
     doc: 'Max concurrency of saveFileForMultihashToFS calls inside nodesync',
     format: 'nat',
     env: 'nodeSyncFileSaveMaxConcurrency',
     default: 10
+  },
+  syncQueueMaxConcurrency: {
+    doc: 'Max concurrency of SyncQueue',
+    format: 'nat',
+    env: 'syncQueueMaxConcurrency',
+    default: 50
   },
 
   // wallet information
@@ -377,7 +383,7 @@ const config = convict({
     doc: 'sync debounce time in ms',
     format: 'nat',
     env: 'debounceTime',
-    default: 30000 // 30000ms = 30s
+    default: 0 // 0ms
   },
   dataRegistryAddress: {
     doc: 'data contracts registry address',
