+++ creator-node/test/syncAssets/README.md
@@ -1,1 +1,3 @@
 `sampleExport.json` was generated against a local dev content node by calling: `http://cn1_creator-node_1:4000/export?wallet_public_key=0x38c8de3ffff2bd062d8557bbbc1cf464b8f199d9`
+
+The values for "multihash" and "storage_path" have been changed to reflect a dummy CID mapping to `Buffer.alloc(32)` to match the mocked `/ipfs` responses in the `nodesync.test.js` file.
